+++ tests/Api/Config/ConfigTest.php
@@ -72,9 +72,9 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         $config = new Config();
         $config->set(Config::PULI_DIR, 'my-puli-dir');
-        $config->set(Config::FACTORY_FILE, '{$puli-dir}/ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_FILE, '{$puli-dir}/ServiceRegistry.php');
 
-        $this->assertSame('{$puli-dir}/ServiceRegistry.php', $config->getRaw(Config::FACTORY_FILE));
+        $this->assertSame('{$puli-dir}/ServiceRegistry.php', $config->getRaw(Config::FACTORY_IN_FILE));
     }
 
     /**
@@ -249,19 +249,19 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         $config = new Config();
         $config->set(Config::PULI_DIR, 'my-puli-dir');
-        $config->set(Config::FACTORY_FILE, '{$puli-dir}/ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_FILE, '{$puli-dir}/ServiceRegistry.php');
 
-        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_FILE));
+        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE));
     }
 
     public function testGetReplacesPlaceholderDefinedInDefaultConfig()
     {
         $baseConfig = new Config();
-        $baseConfig->set(Config::FACTORY_FILE, '{$puli-dir}/ServiceRegistry.php');
+        $baseConfig->set(Config::FACTORY_IN_FILE, '{$puli-dir}/ServiceRegistry.php');
         $config = new Config($baseConfig);
         $config->set(Config::PULI_DIR, 'my-puli-dir');
 
-        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_FILE));
+        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE));
     }
 
     public function testGetReplacesPlaceholderSetInDefaultConfig()
@@ -269,9 +269,9 @@ class ConfigTest extends PHPUnit_Framework_TestCase
         $baseConfig = new Config();
         $baseConfig->set(Config::PULI_DIR, 'my-puli-dir');
         $config = new Config($baseConfig);
-        $config->set(Config::FACTORY_FILE, '{$puli-dir}/ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_FILE, '{$puli-dir}/ServiceRegistry.php');
 
-        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_FILE));
+        $this->assertSame('my-puli-dir/ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE));
     }
 
     public function testGetDoesNotUseFallbackPlaceholderIfFallbackDisabled()
@@ -279,9 +279,9 @@ class ConfigTest extends PHPUnit_Framework_TestCase
         $baseConfig = new Config();
         $baseConfig->set(Config::PULI_DIR, 'my-puli-dir');
         $config = new Config($baseConfig);
-        $config->set(Config::FACTORY_FILE, '{$puli-dir}/ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_FILE, '{$puli-dir}/ServiceRegistry.php');
 
-        $this->assertSame('/ServiceRegistry.php', $config->get(Config::FACTORY_FILE, null, false));
+        $this->assertSame('/ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE, null, false));
     }
 
     /**
@@ -614,26 +614,26 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         $config = new Config();
         $config->set(Config::PULI_DIR, 'puli-dir');
-        $config->set(Config::FACTORY_CLASS, 'Puli\ServiceRegistry');
+        $config->set(Config::FACTORY_IN_CLASS, 'Puli\ServiceRegistry');
         $config->merge(array(
-            Config::FACTORY_CLASS => 'My\ServiceRegistry',
-            Config::FACTORY_FILE => 'repo-file.php',
+            Config::FACTORY_IN_CLASS => 'My\ServiceRegistry',
+            Config::FACTORY_IN_FILE => 'repo-file.php',
         ));
 
         $this->assertSame('puli-dir', $config->get(Config::PULI_DIR));
-        $this->assertSame('My\ServiceRegistry', $config->get(Config::FACTORY_CLASS));
-        $this->assertSame('repo-file.php', $config->get(Config::FACTORY_FILE));
+        $this->assertSame('My\ServiceRegistry', $config->get(Config::FACTORY_IN_CLASS));
+        $this->assertSame('repo-file.php', $config->get(Config::FACTORY_IN_FILE));
     }
 
     public function testRemove()
     {
         $config = new Config();
-        $config->set(Config::FACTORY_FILE, 'ServiceRegistry.php');
-        $config->set(Config::FACTORY_CLASS, 'Puli\ServiceRegistry');
-        $config->remove(Config::FACTORY_CLASS);
+        $config->set(Config::FACTORY_IN_FILE, 'ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_CLASS, 'Puli\ServiceRegistry');
+        $config->remove(Config::FACTORY_IN_CLASS);
 
-        $this->assertSame('ServiceRegistry.php', $config->get(Config::FACTORY_FILE));
-        $this->assertNull($config->get(Config::FACTORY_CLASS));
+        $this->assertSame('ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE));
+        $this->assertNull($config->get(Config::FACTORY_IN_CLASS));
     }
 
     public function testRemoveCompositeKey()
@@ -661,14 +661,14 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     public function testGetReturnsFallbackAfterRemove()
     {
         $baseConfig = new Config();
-        $baseConfig->set(Config::FACTORY_CLASS, 'Fallback\ServiceRegistry');
+        $baseConfig->set(Config::FACTORY_IN_CLASS, 'Fallback\ServiceRegistry');
         $config = new Config($baseConfig);
-        $config->set(Config::FACTORY_FILE, 'ServiceRegistry.php');
-        $config->set(Config::FACTORY_CLASS, 'Puli\ServiceRegistry');
-        $config->remove(Config::FACTORY_CLASS);
+        $config->set(Config::FACTORY_IN_FILE, 'ServiceRegistry.php');
+        $config->set(Config::FACTORY_IN_CLASS, 'Puli\ServiceRegistry');
+        $config->remove(Config::FACTORY_IN_CLASS);
 
-        $this->assertSame('ServiceRegistry.php', $config->get(Config::FACTORY_FILE));
-        $this->assertSame('Fallback\ServiceRegistry', $config->get(Config::FACTORY_CLASS));
+        $this->assertSame('ServiceRegistry.php', $config->get(Config::FACTORY_IN_FILE));
+        $this->assertSame('Fallback\ServiceRegistry', $config->get(Config::FACTORY_IN_CLASS));
     }
 
     public function testToRawArray()
@@ -859,8 +859,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         return array(
             array(Config::PULI_DIR),
-            array(Config::FACTORY_CLASS),
-            array(Config::FACTORY_FILE),
+            array(Config::FACTORY_IN_CLASS),
+            array(Config::FACTORY_IN_FILE),
             array(Config::DISCOVERY_STORE_BUCKET),
         );
     }
@@ -876,8 +876,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         return array(
             array(Config::PULI_DIR),
-            array(Config::FACTORY_CLASS),
-            array(Config::FACTORY_FILE),
+            array(Config::FACTORY_IN_CLASS),
+            array(Config::FACTORY_IN_FILE),
             array(Config::DISCOVERY_STORE_BUCKET),
         );
     }
@@ -886,8 +886,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase
     {
         return array(
             array(Config::PULI_DIR),
-            array(Config::FACTORY_CLASS),
-            array(Config::FACTORY_FILE),
+            array(Config::FACTORY_IN_CLASS),
+            array(Config::FACTORY_IN_FILE),
             array(Config::DISCOVERY_STORE_BUCKET),
         );
     }