+++ tests/Package/RootPackageFileManagerImplTest.php
@@ -79,12 +79,12 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
                 $config = $packageFile->getConfig();
 
                 PHPUnit_Framework_Assert::assertSame('my-puli-dir', $config->get(Config::PULI_DIR));
-                PHPUnit_Framework_Assert::assertSame('my-puli-dir/MyFactory.php', $config->get(Config::FACTORY_FILE));
+                PHPUnit_Framework_Assert::assertSame('my-puli-dir/MyFactory.php', $config->get(Config::FACTORY_IN_FILE));
             }));
 
         $this->manager->setConfigKeys(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ));
     }
 
@@ -103,9 +103,9 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
     public function testGetConfigKeyReturnsRawValue()
     {
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
-        $this->assertSame('{$puli-dir}/MyFactory.php', $this->manager->getConfigKey(Config::FACTORY_FILE));
+        $this->assertSame('{$puli-dir}/MyFactory.php', $this->manager->getConfigKey(Config::FACTORY_IN_FILE));
     }
 
     public function testGetConfigKeyReturnsNullIfNotSet()
@@ -124,56 +124,56 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
 
     public function testGetConfigKeys()
     {
-        $this->baseConfig->set(Config::FACTORY_CLASS, 'My\Class');
+        $this->baseConfig->set(Config::FACTORY_IN_CLASS, 'My\Class');
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
         $this->assertSame(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->getConfigKeys());
     }
 
     public function testGetConfigKeysReordersToDefaultOrder()
     {
-        $this->baseConfig->set(Config::FACTORY_CLASS, 'My\Class');
+        $this->baseConfig->set(Config::FACTORY_IN_CLASS, 'My\Class');
 
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
 
         $this->assertSame(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->getConfigKeys());
     }
 
     public function testGetConfigKeysWithFallback()
     {
-        $this->baseConfig->set(Config::FACTORY_CLASS, 'My\Class');
+        $this->baseConfig->set(Config::FACTORY_IN_CLASS, 'My\Class');
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
         $this->assertSame(array(
             Config::PULI_DIR => 'my-puli-dir',
-            Config::FACTORY_CLASS => 'My\Class',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_CLASS => 'My\Class',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->getConfigKeys(true));
     }
 
     public function testGetConfigKeysWithAllKeys()
     {
-        $this->baseConfig->set(Config::FACTORY_CLASS, 'My\Class');
+        $this->baseConfig->set(Config::FACTORY_IN_CLASS, 'My\Class');
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
         $values = $this->manager->getConfigKeys(false, true);
 
         $this->assertSame(Config::getKeys(), array_keys($values));
         $this->assertSame('my-puli-dir', $values[Config::PULI_DIR]);
-        $this->assertSame('{$puli-dir}/MyFactory.php', $values[Config::FACTORY_FILE]);
+        $this->assertSame('{$puli-dir}/MyFactory.php', $values[Config::FACTORY_IN_FILE]);
         $this->assertNull($values[Config::DISCOVERY_STORE_PATH]);
         $this->assertNull($values[Config::FACTORY_AUTO_GENERATE]);
     }
@@ -183,12 +183,12 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
         $this->baseConfig->set(Config::FACTORY_AUTO_GENERATE, true);
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_CLASS, 'MyFactory');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_CLASS, 'MyFactory');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
         $this->assertSame(array(
-            Config::FACTORY_CLASS => 'MyFactory',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_CLASS => 'MyFactory',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->findConfigKeys('factory.*'));
     }
 
@@ -197,12 +197,12 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
         $this->baseConfig->set(Config::FACTORY_AUTO_GENERATE, true);
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_CLASS, 'MyFactory');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_CLASS, 'MyFactory');
 
         $this->assertSame(array(
-            Config::FACTORY_CLASS => 'MyFactory',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_CLASS => 'MyFactory',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->findConfigKeys('factory.*'));
     }
 
@@ -211,13 +211,13 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
         $this->baseConfig->set(Config::FACTORY_AUTO_GENERATE, true);
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_CLASS, 'MyFactory');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, '{$puli-dir}/MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_CLASS, 'MyFactory');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, '{$puli-dir}/MyFactory.php');
 
         $this->assertSame(array(
             Config::FACTORY_AUTO_GENERATE => true,
-            Config::FACTORY_CLASS => 'MyFactory',
-            Config::FACTORY_FILE => '{$puli-dir}/MyFactory.php',
+            Config::FACTORY_IN_CLASS => 'MyFactory',
+            Config::FACTORY_IN_FILE => '{$puli-dir}/MyFactory.php',
         ), $this->manager->findConfigKeys('factory.*', true));
     }
 
@@ -226,19 +226,21 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
         $this->baseConfig->set(Config::FACTORY_AUTO_GENERATE, true);
 
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_CLASS, 'MyFactory');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_CLASS, 'MyFactory');
 
         $this->assertSame(array(
             Config::FACTORY_AUTO_GENERATE => true,
-            Config::FACTORY_CLASS => 'MyFactory',
-            Config::FACTORY_FILE => null,
+            Config::FACTORY_IN_CLASS => 'MyFactory',
+            Config::FACTORY_IN_FILE => null,
+            Config::FACTORY_OUT_CLASS => null,
+            Config::FACTORY_OUT_FILE => null,
         ), $this->manager->findConfigKeys('factory.*', true, true));
     }
 
     public function testRemoveConfigKey()
     {
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, 'MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, 'MyFactory.php');
 
         $this->packageFileStorage->expects($this->once())
             ->method('saveRootPackageFile')
@@ -247,7 +249,7 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
                 $config = $packageFile->getConfig();
 
                 PHPUnit_Framework_Assert::assertNull($config->get(Config::PULI_DIR, null, false));
-                PHPUnit_Framework_Assert::assertSame('MyFactory.php', $config->get(Config::FACTORY_FILE, null, false));
+                PHPUnit_Framework_Assert::assertSame('MyFactory.php', $config->get(Config::FACTORY_IN_FILE, null, false));
             }));
 
         $this->manager->removeConfigKey(Config::PULI_DIR);
@@ -256,7 +258,7 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
     public function testRemoveConfigKeys()
     {
         $this->rootPackageFile->getConfig()->set(Config::PULI_DIR, 'my-puli-dir');
-        $this->rootPackageFile->getConfig()->set(Config::FACTORY_FILE, 'MyFactory.php');
+        $this->rootPackageFile->getConfig()->set(Config::FACTORY_IN_FILE, 'MyFactory.php');
 
         $this->packageFileStorage->expects($this->once())
             ->method('saveRootPackageFile')
@@ -265,10 +267,10 @@ class RootPackageFileManagerImplTest extends ManagerTestCase
                 $config = $packageFile->getConfig();
 
                 PHPUnit_Framework_Assert::assertNull($config->get(Config::PULI_DIR, null, false));
-                PHPUnit_Framework_Assert::assertNull($config->get(Config::FACTORY_FILE, null, false));
+                PHPUnit_Framework_Assert::assertNull($config->get(Config::FACTORY_IN_FILE, null, false));
             }));
 
-        $this->manager->removeConfigKeys(array(Config::PULI_DIR, Config::FACTORY_FILE));
+        $this->manager->removeConfigKeys(array(Config::PULI_DIR, Config::FACTORY_IN_FILE));
     }
 
     public function testAddPluginClass()